+++ src/hal/ARContractSearch.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace roaresearch\yii2\roa\hal;
+
+use roaresearch\yii2\roa\ResourceSearch;
+
+interface ARContractSearch extends ARContract, ResourceSearch
+{
+}