+++ src/urlRules/Modular.php
@@ -19,18 +19,25 @@ class Modular extends Composite
      *
      * - {moduleId}: the unique module id associated to this rule.
      */
-    public $notFoundMessage = 'Unknown route for module `{moduleId}`.';
+    public string $notFoundMessage = 'Unknown route for module `{moduleId}`.';
 
     /**
      * @var string unique id to grab the module from the application that will
      * parse the rules.
      */
-    public $moduleId;
+    public string $moduleId;
 
     /**
      * @var UrlRuleCreator
      */
-    protected $module;
+    protected UrlRuleCreator $module;
+
+    /**
+     * @inheritdoc
+     */    
+    public function init()
+    {
+    }
 
     /**
      * @inheritdoc
@@ -62,16 +69,7 @@ class Modular extends Composite
 
     protected function ensureRules()
     {
-        if (empty($this->moduleId)) {
-            throw new InvalidConfigException('`moduleId` must be set.');
-        }
         $this->module = Yii::$app->getModule($this->moduleId);
-        if (!$this->module instanceof UrlRuleCreator) {
-            throw new InvalidConfigException(
-                "Module `{$this->moduleId}` must implement "
-                    . UrlRuleCreator::class
-            );
-        }
         $this->module->initCreator($this);
 
         parent::ensureRules();