+++ tests/_app/api/models/Item.php
@@ -2,13 +2,13 @@
 
 namespace app\api\models;
 
-use roaresearch\yii2\roa\hal\{Contract, ContractTrait};
-use yii\web\NotFoundHttpException;
+use roaresearch\yii2\roa\{behaviors\Slug, hal\ARContract, hal\ContractTrait};
+use yii\{base\Action, web\NotFoundHttpException};
 
 /**
  * ROA contract to handle item records.
  */
-class Item extends \app\models\Item implements Contract
+class Item extends \app\models\Item implements ARContract
 {
     use ContractTrait {
         getLinks as getContractLinks;
@@ -17,25 +17,30 @@ class Item extends \app\models\Item implements Contract
     /**
      * @inheritdoc
      */
-    protected $employeeClass = Employee::class;
+    protected string $employeeClass = Employee::class;
 
     /**
      * @inheritdoc
      */
-    protected function slugBehaviorConfig(): array
+    protected function slugBehaviorConfig(): Slug
     {
-        return [
-            'resourceName' => 'item',
-            'checkAccess' => function ($params) {
-                if (isset($params['item_id'])
-                    && $this->id != $params['item_id']
+        return new class (['owner' => $this]) extends Slug {
+            public string $resourceName = 'item';
+
+            public function checkAccess(
+                array $params = [],
+                ?Action $action = null
+            ): void {
+                if (
+                    isset($params['item_id'])
+                    && $this->owner->id != $params['item_id']
                 ) {
                     throw new NotFoundHttpException(
-                       'Item not associated to element.'
+                        'Item not associated to element.'
                     );
                 }
-            },
-        ];
+            }
+        };
     }
 
     /**