+++ tests/_app/api/models/Shop.php
@@ -2,13 +2,13 @@
 
 namespace app\api\models;
 
-use roaresearch\yii2\roa\hal\{Contract, ContractTrait};
-use yii\web\NotFoundHttpException;
+use roaresearch\yii2\roa\{behaviors\Slug, hal\ARContract, hal\ContractTrait};
+use yii\{base\Action, web\NotFoundHttpException};
 
 /**
  * ROA contract to handle shop records.
  */
-class Shop extends \app\models\Shop implements Contract
+class Shop extends \app\models\Shop implements ARContract
 {
     use ContractTrait {
         getLinks as getContractLinks;
@@ -17,25 +17,30 @@ class Shop extends \app\models\Shop implements Contract
     /**
      * @inheritdoc
      */
-    protected $employeeClass = Employee::class;
+    protected string $employeeClass = Employee::class;
 
     /**
      * @inheritdoc
      */
-    protected function slugBehaviorConfig(): array
+    protected function slugBehaviorConfig(): Slug
     {
-        return [
-            'resourceName' => 'shop',
-            'checkAccess' => function ($params) {
-                if (isset($params['shop_id'])
-                    && $this->id != $params['shop_id']
+        return new class (['owner' => $this]) extends Slug {
+            public string $resourceName = 'shop';
+
+            public function checkAccess(
+                array $params = [],
+                ?Action $action = null
+            ): void {
+                if (
+                    isset($params['shop_id'])
+                    && $this->owner->id != $params['shop_id']
                 ) {
                     throw new NotFoundHttpException(
-                       'Shop not associated to element.'
+                        'Shop not associated to element.'
                     );
                 }
-            },
-        ];
+            }
+        };
     }
 
     /**
@@ -46,6 +51,14 @@ class Shop extends \app\models\Shop implements Contract
         return array_merge($this->getContractLinks(), [
             'employee' => $this->getSelfLink() . '/employee',
         ]);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function fields()
+    {
+        return ['id', 'name'];
     }
 
     /**